requirements.txt
fastapi==0.112.2
matplotlib==3.6.3
moviepy==1.0.3
numpy<2
Pillow==10.4.0
pydub==0.25.1
scipy==1.14.1
uvicorn==0.30.6
python-multipart==0.0.9
audioop-lts==0.2.2